+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Adds oihana\core\accessors\hasKeyValue
 - Adds oihana\core\accessors\setKeyValue
 - Adds oihana\core\date\formatDateTime
+- Adds oihana\core\documents\formatDocument
 - Adds oihana\core\reflections\getFunctionInfo
 - Adds oihana\core\strings\block
 - Adds oihana\core\strings\blockPrefix